add prompt before delete a customer

+++ WebContent/WEB-INF/view/list-customers.jsp
@@ -51,7 +51,7 @@
 						<td> ${tempCustomer.firstName } </td>
 						<td> ${tempCustomer.lastName } </td>
 						<td> ${tempCustomer.email } </td>
-						<td> <a href="${updateLink}">Update</a>| <a href="${deleteLink}">Delete</a></td>
+						<td> <a href="${updateLink}">Update</a>| <a href="${deleteLink}" onclick="if (!(confirm('Are you sure you want to delete this customer?'))) return false">Delete</a></td>
 
 					</tr>
